refactor: Height 클래스 패키지 이동 및 레코드로 수정
