+++ src/main/java/ladder/domain/ladder/size/Height.java
@@ -1,26 +1,17 @@
-package ladder.domain.ladder;
+package ladder.domain.ladder.size;
 
 import ladder.exception.ladder.InvalidHeightRangeException;
 
-import java.util.List;
-
-public class Height {
+public record Height(int value) {
     private static final int MIN_HEIGHT = 1;
 
-    private final int value;
-
-    public Height(final int value) {
+    public Height {
         validateHeightRange(value);
-        this.value = value;
     }
 
     private void validateHeightRange(final int value) {
         if (value < MIN_HEIGHT) {
             throw new InvalidHeightRangeException();
         }
-    }
-
-    public boolean isGreaterThan(final List<LadderStep> ladderSteps) {
-        return value > ladderSteps.size();
     }
 }
