+++ src/test/java/ladder/domain/ladder/HeightTest.java
@@ -1,5 +1,6 @@
 package ladder.domain.ladder;
 
+import ladder.domain.ladder.size.Height;
 import ladder.exception.ladder.InvalidHeightRangeException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
